Update ip to make it work as container
We need host="0.0.0.0". This is a necessity as 0.0.0.0 is a wildcard IP address that will match any possible incoming port on the host machine. We require this because the default port 127.0.0.1 inside a Docker container does not actually get captured and exposed into the host machine.

Also avoid debug mode